Run system:msvc:windows:amd64 for merge requests
Historically, we avoided running system tests on Windows for merge
request-triggered pipelines because we have only one limited GitLab
Runner, which would be unable to reliably sustain the CI load of merge
requests for three maintained branches, and instead, we opted for
running Windows system tests in daily pipelines. CI load should no
longer be an issue since "v9_16" is the only (and last) maintained
branch with Windows support. We are only left with the downside of
failing to run system tests for merge requests on Windows: missing
Windows-specific quirks in the review process and finding out about them
only a day later in the daily pipeline (e.g., #3865 and #3874 from this
week).